+++ modules/orionode.collab.hub/config.js
@@ -13,11 +13,11 @@ module.exports = {
 	/**
 	 * jwt_secret needs to be the same as the one in the server connected to the filesystem (orion)
 	 */
-	'jwt_secret': "pomato (potato and tomato mix lol)",
+	'jwt_secret': "orion collab",
 	/**
 	 * Orion url
 	 */
-	'orion': "http://localhost:8081/",
+	'orion': "http://localhost:8084/",
 	/**
 	 * Load url. Make sure end with /
 	 */
